make file visualizations work
