+++ molecules_binding/__init__.py
@@ -3,6 +3,5 @@
 from .models import *
 from .parsers import *
 from .lightning_wrapper import *
-from .visualization_functions import *
 from .datasets_interaction import *
 from .parsers_interaction import *
